Break ViewNotesCommandTest into Logic test and UI test for better test segmentation
