+++ src/test/java/seedu/address/logic/commands/ViewNotesCommandLogicTest.java
@@ -1,36 +1,21 @@
 package seedu.address.logic.commands;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 import static seedu.address.testutil.Assert.assertThrows;
 
 import org.junit.jupiter.api.Test;
-import org.testfx.framework.junit5.ApplicationTest;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import javafx.stage.Stage;
+
 import seedu.address.logic.Messages;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.person.Person;
 import seedu.address.testutil.ModelStub;
 import seedu.address.testutil.PersonBuilder;
 
-public class ViewNotesCommandTest extends ApplicationTest {
-
-    @Override
-    public void start(Stage stage) {};
-
-    @Test
-    public void execute_viewNotes_successful() throws Exception {
-        Person personA = new PersonBuilder().withName("A").build();
-        ModelStubWithPerson modelStub = new ModelStubWithPerson(personA);
-        CommandResult commandResult = new ViewNotesCommand(0).execute(modelStub);
-        assertEquals(String.format(ViewNotesCommand.MESSAGE_VIEW_NOTES_SUCCESS,
-            personA.getName()), commandResult.getFeedbackToUser());
-    }
-
+public class ViewNotesCommandLogicTest {
     @Test
     public void execute_invalidIndex_throwsCommandException() {
         Person personA = new PersonBuilder().withName("A").build();
